requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
requests>=2.32.0
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.6.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3
pymupdf>=1.23.0
lxml>=4.9.0
zeep>=4.2.0
reportlab>=4.0.0
opencv-python>=4.8.0
numpy>=1.24.0
redis>=5.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

